# gen/http/openapi3.yaml
openapi: 3.0.3
info:
  title: goa_starterulator Service
  description: Service for adding numbers, a Goa teaser
  version: "1.0"
servers:
- url: http://localhost:8085/api/v1
paths:
  /add/{a}/{b}:
    get:
      tags:
      - goa_starter
      summary: add goa_starter
      operationId: goa_starter#add
      parameters:
      - name: a
        in: path
        description: Left operand
        required: true
        schema:
          type: integer
          description: Left operand
          example: 8034677293779011882
          format: int64
        example: 36983429280391996
      - name: b
        in: path
        description: Right operand
        required: true
        schema:
          type: integer
          description: Right operand
          example: 1553051376274246529
          format: int64
        example: 2598324719697404761
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                type: integer
                example: 7305426396677949776
                format: int64
              example: 7905194824815223204
  /term_limit/{c}/{d}:
    get:
      tags:
      - term_limit
      summary: add term_limit
      operationId: term_limit#add
      parameters:
      - name: c
        in: path
        description: Left operand
        required: true
        schema:
          type: integer
          description: Left operand
          example: 3481953999106781257
          format: int64
        example: 1253733055766829795
      - name: d
        in: path
        description: Right operand
        required: true
        schema:
          type: integer
          description: Right operand
          example: 6782716350968002240
          format: int64
        example: 8024987327736100917
      responses:
        "200":
          description: OK response.
          content:
            application/json:
              schema:
                type: integer
                example: 7241277749342655831
                format: int64
              example: 8969694167088891872
components: {}
tags:
- name: goa_starter
  description: The goa_starter service performs operations on numbers.
- name: term_limit
  description: The term_limit service performs operations on numbers.